refactor: move my-units page
