+++ apps/admin-ui/src/component/my-units/MyUnitRecurringReservation/hooks.test.tsx
@@ -3,11 +3,11 @@ import { render, waitFor } from "@testing-library/react";
 import { MockedProvider } from "@apollo/client/testing";
 import { add, format, set } from "date-fns";
 import userEvent from "@testing-library/user-event";
-import { RecurringReservationForm } from "app/schemas";
-import { NewReservationListItem } from "app/component/ReservationsList";
+import type { RecurringReservationForm } from "@/schemas";
+import type { NewReservationListItem } from "@component/ReservationsList";
 import { useCreateRecurringReservation } from "./hooks";
 import { mocks } from "./__test__/mocks";
-import { ReservationMade } from "./RecurringReservationDone";
+import type { ReservationMade } from "../schemas";
 import { CREATE_RECURRING_RESERVATION } from "./queries";
 import { CREATE_STAFF_RESERVATION } from "../create-reservation/queries";
