+++ apps/admin-ui/src/component/reservations/requested/RequestedReservation.tsx
@@ -38,7 +38,7 @@ import {
 import Calendar from "./Calendar";
 import VisibleIfPermission from "./VisibleIfPermission";
 import ApprovalButtons from "./ApprovalButtons";
-import RecurringReservationsView from "./RecurringReservationsView";
+import { RecurringReservationsView } from "./RecurringReservationsView";
 import { useRecurringReservations, useReservationData } from "./hooks";
 import ApprovalButtonsRecurring from "./ApprovalButtonsRecurring";
 import ReservationTitleSection from "./ReservationTitleSection";
