+++ apps/admin-ui/src/spa/my-units/MyUnitsRouter.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 import { Route, Routes } from "react-router-dom";
-import MyUnitRecurringReservation from "./MyUnitRecurringReservation/MyUnitRecurringReservation";
-import MyUnits from "./MyUnits";
-import MyUnitView from "./MyUnitView";
-import RecurringReservationDone from "./MyUnitRecurringReservation/RecurringReservationDone";
+import MyUnitRecurringReservation from "./[id]/recurring";
+import MyUnits from "./index";
+import MyUnitView from "./[id]";
+import RecurringReservationDone from "./[id]/recurring/completed";
 
-const MyUnitsRouter = (): JSX.Element => (
+export const MyUnitsRouter = (): JSX.Element => (
   <Routes>
     <Route index element={<MyUnits />} />
     <Route path=":unitId" element={<MyUnitView />} />
@@ -19,5 +19,3 @@ const MyUnitsRouter = (): JSX.Element => (
     />
   </Routes>
 );
-
-export default MyUnitsRouter;
